+++ karma-test-shim.js
@@ -45,7 +45,6 @@ System.config({
         "@ng2-dynamic-forms/ui-foundation": "npm:@ng2-dynamic-forms/ui-foundation/bundles/ui-foundation.umd.js",
         "@ng2-dynamic-forms/ui-material": "npm:@ng2-dynamic-forms/ui-material/bundles/ui-material.umd.js",
         "@ng2-dynamic-forms/ui-primeng": "npm:@ng2-dynamic-forms/ui-primeng/bundles/ui-primeng.umd.js",
-        "angular2-in-memory-web-api": "npm:angular2-in-memory-web-api",
         "primeng": "npm:primeng",
         "rxjs": "npm:rxjs"
     },