+++ modules/core/src/model/select/dynamic-select.model.spec.ts
@@ -4,15 +4,27 @@ describe("DynamicSelectModel test suite", () => {
 
     describe("default model test suite", () => {
 
-        let config = {id: "default", options: [{value: "1"}, {value: "2"}]};
-        let defaultModel: DynamicSelectModel<any>;
+        let defaultModel: DynamicSelectModel<string>,
+            config = {
+                id: "default",
+                options: [
+                    {
+                        value: "1",
+                        label: "One"
+                    },
+                    {
+                        value: "2",
+                        label: "Two"
+                    }
+                ]
+            };
+
 
         beforeEach(() => {
             defaultModel = new DynamicSelectModel(config);
         });
 
         it("tests if default model is correctly initialized", () => {
-
 
             expect(defaultModel.disabled).toBe(false);
             expect(defaultModel.id).toEqual(config.id);
@@ -21,6 +33,27 @@ describe("DynamicSelectModel test suite", () => {
             expect(defaultModel.options.length).toBe(config.options.length);
             expect(defaultModel.type).toEqual(DYNAMIC_FORM_CONTROL_TYPE_SELECT);
             expect(defaultModel.value).toBeNull();
+        });
+
+        it("should get and set text property correctly", () => {
+
+            expect(defaultModel.options[0].text).toEqual("One");
+
+            defaultModel.options[0].text = "Eins";
+
+            expect(defaultModel.options[0].text).toEqual("Eins");
+        });
+
+        it("should get the correct option", () => {
+
+            expect(defaultModel.get(0)).toEqual(defaultModel.options[0]);
+            expect(defaultModel.get(1)).toEqual(defaultModel.options[1]);
+        });
+
+        it("should select correct option", () => {
+
+            defaultModel.select(1);
+            expect(defaultModel.value).toEqual(defaultModel.options[1].value);
         });
 
         it("should serialize correctly", () => {