+++ modules/ui-material/src/dynamic-form-material.component.ts
@@ -1,8 +1,7 @@
 import {Component, Input, Output, EventEmitter, ViewChild, ContentChild, TemplateRef} from "@angular/core";
 import {FormGroup} from "@angular/forms";
 import {MdCheckbox, MdInput, MdRadioGroup} from "@angular/material";
-import {DynamicFormControlModel} from "@ng2-dynamic-forms/core";
-import {DynamicFormControlComponent} from "@ng2-dynamic-forms/core";
+import {DynamicFormControlComponent, DynamicFormControlModel, DynamicFormControlEvent} from "@ng2-dynamic-forms/core";
 
 export const DYNAMIC_FORM_UI_MATERIAL = "MATERIAL";
 
@@ -20,9 +19,9 @@ export class DynamicFormMaterialComponent extends DynamicFormControlComponent {
     @Input() model: DynamicFormControlModel;
     @Input() nestedTemplate: TemplateRef<any>;
 
-    @Output() blur: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
-    @Output() change: EventEmitter<Event> = new EventEmitter<Event>();
-    @Output() focus: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
+    @Output() blur: EventEmitter<DynamicFormControlEvent> = new EventEmitter<DynamicFormControlEvent>();
+    @Output() change: EventEmitter<DynamicFormControlEvent> = new EventEmitter<DynamicFormControlEvent>();
+    @Output() focus: EventEmitter<DynamicFormControlEvent> = new EventEmitter<DynamicFormControlEvent>();
 
     @ContentChild(TemplateRef) customTemplate;