+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# 1.2.0
+
+### **Breaking Changes** 
+
+* `DynamicFormControlComponent` **does now emit a** `DynamicFormControlEvent` **on blur, change and focus instead of simply passing through** `$event`:
+ ```
+ export interface DynamicFormControlEvent {
+     
+     $event: Event;
+     control: FormControl;
+     model: DynamicFormControlModel;
+ }
+ ```
+* **blur, change and focus events are only emitted for single form controls and checkbox groups from now on** (see Bootstrap example)
+
+### **Features** 
+
+* updated to `@angular 2.2`
+* `valueUpdates Rx.Observable` can now safely be used from anywhere to listen to new values
+* `checked` setter/getter and `toggle()` function added to `DynamicCheckControlModel`
+* `checkAll()` and `uncheckAll()` function added to `DynamicCheckboxGroupModel`
+* `select()` and `get()` function added to `DynamicOptionControlModel`
+* `get()`function added to `DynamicFormGroupModel` and `DynamicFormArrayModel`
+* textarea support for `ui-material` added
+
+
 # 1.1.3
 
 ### **Bugfixes**